Moves away from any prints or custom loggers in favor of the builtin.
Don't have to track who/where it was logged as its done automatically and now
by default a pybert.log file is created in the current working directory.  It is noisy
but provides a lot of information such as line/time/level/exception trace and etc...

You can still prompt a user in the GUI just like before.  Instead of log.info("message"),
you just have to use log.info("message", {"alert":True}).  This gets passed to the LogRecord
as a kwarg and later handled by the new ConsoleTextLogHandler.